[chore] move authentication to a module
