+++ bucketlist_api/api.py
@@ -13,29 +13,10 @@ from bucketlist_api.models import User, BucketList, BucketListItem
 from bucketlist_api import create_app, api
 from bucketlist_api.config import DevConfig
 from datetime import datetime
+from bucketlist_api.authentication import auth
 
 # initialization
 app = create_app(DevConfig)
-
-# Authentication Object
-auth = HTTPBasicAuth()
-
-
-@auth.verify_password
-def authenticate_token(token, password):
-    """Autheticate User with the provideded token."""
-    if User.verify_token(token):
-        return True
-    return False
-
-
-@auth.error_handler
-def unauthorize():
-    """Handles unauthorize access to the API."""
-
-    return make_response(jsonify({'Error': 'Invalid token Supplied or token '
-                                  'has expired, Login again to get access'
-                                  ' token'}), 401)
 
 
 class CreateUserAPI(Resource):
